refactor: files structure and typing
